LayoutTests/fast/js/kde/math-expected.txt

PASS isNegativeZero(negativeZero) is true
PASS isNegativeZero(0) is false
PASS String()+Math.E is '2.718281828459045'
PASS String()+Math.LN2 is '0.6931471805599453'
PASS String()+Math.LN10 is '2.302585092994046'
PASS String()+Math.LOG2E is '1.4426950408889634'
PASS String()+Math.LOG10E is '0.43429448190325176'
PASS String()+Math.PI is '3.141592653589793'
PASS String()+Math.SQRT1_2 is '0.7071067811865476'
PASS String()+Math.SQRT2 is '1.4142135623730951'
PASS String()+Number.NaN is 'NaN'
PASS String()+Number.NEGATIVE_INFINITY is '-Infinity'
PASS String()+Number.POSITIVE_INFINITY is 'Infinity'
PASS Math.abs(-5) is 5
PASS Math.acos(0) is Math.PI/2
PASS Math.acos(1) is 0
PASS Math.ceil(1.1) is 2
PASS String()+Math.sqrt(2) is String()+Math.SQRT2
PASS Math.ceil(1.6) is 2
PASS Math.round(0) is 0
PASS isNegativeZero(Math.round(0)) is false
PASS isNegativeZero(Math.round(negativeZero)) is true
PASS Math.round(0.2) is 0
PASS isNegativeZero(Math.round(-0.2)) is true
PASS isNegativeZero(Math.round(-0.5)) is true
PASS Math.round(1.1) is 1
PASS Math.round(1.6) is 2
PASS Math.round(-3.5) is -3
PASS Math.round(-3.6) is -4
PASS isNaN(Math.round()) is true
PASS isNaN(Math.round(NaN)) is true
PASS Math.round(-Infinity) is -Infinity
PASS Math.round(Infinity) is Infinity
PASS Math.round(99999999999999999999.99) is 100000000000000000000
PASS Math.round(-99999999999999999999.99) is -100000000000000000000
PASS Math.log(Math.E*Math.E) is 2
PASS isNaN(Math.log(NaN)) is true
PASS isNaN(Math.log(-1)) is true
PASS isFinite(Math.log(0)) is false
PASS Math.log(1) is 0
PASS isFinite(Math.log(Infinity)) is false
PASS isNegativeZero(Math.min(negativeZero, 0)) is true
PASS isFinite(Math.max()) is false
PASS Math.max(1) is 1
PASS Math.max(1, 2, 3) is 3
PASS isNaN(Math.max(1,NaN,3)) is true
PASS !isNegativeZero(Math.max(negativeZero, 0)) is true
PASS list is ''
PASS delete my.v is true
PASS my.v is undefined.
PASS delete Math.PI is false
PASS my = myfunc, myfunc(4) is 5
PASS Boolean(Math) is true
PASS isNaN(Number(Math)); is true
PASS Math.abs===Math.abs is true
PASS Math.abs===Math.round is false
PASS list is 'a,b,'
PASS list is ''
PASS list is 'myprop,'
PASS successfullyParsed is true
